Re-enable USB gadget init.rc for OnePlus7
The init.nethunter.rc used to copy a keyboard and mouse descriptor file for
HID gadget mode. Copying these files causes boot loops on OnePlus 7
devices and the files are not needed. The system takes care of that
already. Removed these lines and fixed some spelling mistakes.

TODO: Both descriptor files are still left as reference for a current
change to usbarmory and should be removed as soon as usbarmory is
updated.

Signed-off-by: Re4son <3520622-re4son@users.noreply.gitlab.com>